refactor(camelot): rename test-transactions.ts to test-setup.ts
